feat: Use perl instead of awk
On my system (Debian) awk is actually mawk, which does not support
word-boundaries in regular expressions, and makes searching for exact
tags much more annoying.

Another option would be to modify the format to make all tags have a
prefix and suffix, e.g. `:book:` instead of `book`, but that's more
burden on the user.